simulator/requirements.txt
fastapi
uvicorn[standard]
kafka-python
pandas
joblib
prometheus-client
scikit-learn>=1.0